Update prep-media script, only copy if contents are different

+++ PrepareMedia.py
@@ -5,6 +5,7 @@ import os
 import sys 
 import string
 import shutil
+import filecmp
 
 inBuildDir = None
 inMediaDir = None
@@ -36,7 +37,9 @@ def CopyFile(sourceDir : string, destinationDir : string):
     if not os.path.exists(destinationDir):
         os.makedirs(destinationDir)
     finalDestination = os.path.join(destinationDir, fileName)
-    shutil.copy2(sourceDir, finalDestination)
+
+    if os.path.isfile(finalDestination) and not filecmp.cmp(sourceDir, finalDestination):
+        shutil.copy2(sourceDir, finalDestination)
 
 #copy media as needed 
 def SearchForFiles(currentSourcePath : string, currentDestPath : string):
